requirements.txt
# Statistical Arbitrage Platform Requirements
# ==========================================

# Core Data Science
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
statsmodels>=0.14.0

# Machine Learning & Statistics
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Async HTTP & WebSocket
aiohttp>=3.8.0
websockets>=11.0
httpx>=0.24.0

# Configuration & Data
yaml>=0.2.5
pyyaml>=6.0
python-dotenv>=1.0.0

# Database & Storage
sqlalchemy>=2.0.0
pandas-datareader>=0.10.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0

# Development Tools
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0
flake8>=6.0.0

# API & Data Sources
yfinance>=0.2.20
ccxt>=4.0.0
requests>=2.31.0

# Monitoring & Logging
structlog>=23.1.0
prometheus-client>=0.17.0

# Mathematical & Financial
QuantLib-Python>=1.31
pykalman>=0.9.5

# Time Series
arch>=6.2.0  # For GARCH models

# Optional: Advanced Features
# redis>=4.6.0  # For caching
# celery>=5.3.0  # For task queues
# fastapi>=0.100.0  # For web API
# uvicorn>=0.23.0  # ASGI server

# Cryptocurrency APIs
python-binance>=1.0.17

# Risk & Portfolio Management
empyrical>=0.5.5
pyfolio>=0.9.2

# Data Validation
pydantic>=2.0.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0